backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose==3.3.0
